Triangulation: Move definitions of functions to traingulation.cpp
